pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wallpaper_automator" 
version = "0.1.0"
description = "Automatically switches Windows desktop wallpapers based on configurable conditions (time, WiFi, location, workday)"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Mike", email = "houlin.mike@gmail.com" }
]
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Environment :: Win32 (MS Windows)",
]

dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pywin32>=306",
    "Pillow>=10.0",
    "requests>=2.28",
    "fake_useragent>=1.4",
    "wmi>=1.5",
    "tenacity>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "ruff>=0.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q --tb=short --cov=src"
python_files = "test_*.py"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "**/__init__.py", "**/__main__.py"]

[tool.coverage.report]
show_missing = true
precision = 2

[tool.mypy]
python_version = "3.9"
strict = true
ignore_missing_imports = true
disallow_untyped_defs = true
warn_unused_ignores = true
exclude = ["tests/"]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.format]
quote-style = "double"
